Fix for deletion of file content upon calling the script
Every time a call to .setValue() or .get(), the line open(folder,"w").write("{}") would get executed, thus clearing the file of all stored variables. The script now checks for the file existence, and creates it if it does not.

+++ guli/guli.py
@@ -1,7 +1,10 @@
 import json
 import os
+from os.path import exists
+
 folder = os.path.dirname(os.path.realpath(__file__))+"/data.json"
-open(folder,"w").write("{}")
+if (exists(folder) == False):
+  open(folder,"w").write("{}")
 
 def GuliVariable(variable_name):
   return __global__(variable_name)
